Fix: ExpressionSimplifier
* Avoid mutable state in ExpressionSimplifier by changing its methods' return types to (Expression, bool), where the bool indicates "changed".
* New slice analyses to reduce the number of casts and SLICE expressions in the final output. (#1072)

+++ subjects/PE/MIPS/swlswr/test.reko/test_text.c
@@ -83,7 +83,7 @@ void fn00011278(Eq_n r4)
 //      fn00011434
 void fn000112A8(Eq_n r4, int32 r5, int32 r6)
 {
-	g_b13038 = (byte) (int32) (byte) r6;
+	g_b13038 = (byte) r6;
 	if (r5 != 0x00)
 	{
 		fn0001147C(&g_dw12018, &g_dw1201C);
